user-story-tests/27-markdown-component-with-automatic-page-style.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Markdown component with automatic page styles</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <script type="module">
        import { MDBlock } from "./components/MDBlock.js";
        customElements.define(
            "md-block",
            MDBlock
        );
    </script>

    <style>
        h2 {
            border: thick dashed red;
        }
    </style>
    <h2>H2 outside shadow tree (thick dashed red)</h2>
    <md-block>
        <script type="text/markdown">
## H2 from markdown (thick dashed red) (from page styles)
        </script>
    </md-block>

    <script type="module">
        import { adoptStylesFromLayerStatementRule } from "../AdoptStyles.js";

        window.onload = function () {

            test(function () {

                assert_true(true, "always true, just to show the test is running");

            }, "Markdown component with automatic page styles");

        }
    </script>

    <div id="log"></div>

</body>

</html>
